separate schema files, move computer
