+++ qcmanybody/models/__init__.py
@@ -1,8 +1,12 @@
-from .manybody_pydv1 import (
+from .manybody_input_pydv1 import (
+    AtomicSpecification,
     BsseEnum,
     FragBasIndex,
     ManyBodyInput,
     ManyBodyKeywords,
+)
+from .manybody_output_pydv1 import (
     ManyBodyResult,
     ManyBodyResultProperties,
+    MAX_NBODY
 )
